e2e-tests/pytest/pyproject.toml
[project]
name = "screw"
version = "0.1.0"
description = ""
authors = [
    {name = "kubescrew", email = "cloud@percona.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
	"kubernetes==32.0.0",
	"pyyaml (>=6.0.2,<7.0.0)",
	"google-cloud-container (>=2.55.1,<3.0.0)",
	"google-auth (>=2.38.0,<3.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
